remove next-mdx-remote
