+++ package.json
@@ -33,7 +33,6 @@
     "lodash": "^4.17.21",
     "next": "^13.2.4",
     "next-auth": "^4.21.0",
-    "next-mdx-remote": "^4.4.1",
     "prism-react-renderer": "^2.0.3",
     "react": "18.2.0",
     "react-content-loader": "^6.2.1",
