+++ src/components/MarkdownRenderer/index.ts
@@ -1,2 +1,1 @@
-export * from './components/MarkdocRenderer'
 export * from './components/MarkdownRenderer'
