+++ src/pages/posts/[id].tsx
@@ -1,5 +1,5 @@
 import { MainLayout } from '@/components/Layout'
-import { MarkdocRenderer } from '@/components/MarkdownRenderer'
+import { MarkdownRenderer } from '@/components/MarkdownRenderer'
 import { prisma } from '@/server/prisma'
 import { type GetStaticPaths, type GetStaticProps, type NextPage } from 'next'
 
@@ -64,7 +64,7 @@ const PostPage: NextPage<Props> = ({ post }) => {
       <h1 className="mb-4 inline-block text-4xl font-extrabold leading-tight text-slate-900 lg:text-5xl">
         {post.title}
       </h1>
-      {<MarkdocRenderer content={post.content} />}
+      <MarkdownRenderer content={post.content} />
     </MainLayout>
   )
 }
